Spanish translation of the site is working
Although yeah the site now has a Spanish version its not the best, nor as dynamic I wish for it to be. I wanted to make a JSON file that would dynamically change the language whenever the user feels like it. Although I know how, I do not want to create an unintentional blow-up of other features. So, instead I would like to maybe remake the site in React app just so it makes more sense for me... and more practice for myself.