Some accessibility improvements
This is something new the AI was teaching me about. Granted I do not normally use these but I found them interesting enough to possibly help people if actually useful. So, yes this is ai, I dont like that I didnt write much of this and my testing is a little limited in practical use so, I will holp to continue testing later